Update
API Update

Summoner now optimised for lower levels ( UCOB,UWU,TEA should be performing better ):
Added safety to avoid using swifcast for lower level when Garuda to avoid loosing the swift
Also Changed the Radiant aegis to be used even if only one stack is available on lower level to keep the function "Use Radiant On Cooldown" consistent.
And optimized some of the basic stuff which made the tincture usage better timed.

Added modified rotation of RedMage
STILL NEED INTENSIVE TESTING
However, the tincture should be a bit better timed. Manafication should be used after a first rotation when bursting if the player begin his burst with already enough black and white magic to perform a combo. Some times the rotation was making mistakes if you pushed the settings of timings a bit low and have no animation lock. I also modified this to be avoided should not be occuring anymore.

The basic of this rotation was made BY ArchiTed, so still show big love to him he's the boss !